Split out TopBarIconButtonV2 file + TopBarIconButtonV2.test file
