+++ src/navigator/Headers.tsx
@@ -18,7 +18,8 @@ import AccountCircle from 'src/icons/AccountCircle'
 import BackChevronCentered from 'src/icons/BackChevronCentered'
 import { navigate, navigateBack } from 'src/navigator/NavigationService'
 import { Screens } from 'src/navigator/Screens'
-import { TopBarIconButton, TopBarIconButtonV2 } from 'src/navigator/TopBarButton'
+import { TopBarIconButton } from 'src/navigator/TopBarButton'
+import { TopBarIconButtonV2 } from 'src/navigator/TopBarIconButtonV2'
 import DisconnectBanner from 'src/shared/DisconnectBanner'
 import { getFeatureGate } from 'src/statsig'
 import { StatsigFeatureGates } from 'src/statsig/types'
