+++ src/navigator/TopBarButton.test.tsx
@@ -1,7 +1,7 @@
 import { fireEvent, render } from '@testing-library/react-native'
 import * as React from 'react'
 import { Text } from 'react-native'
-import { TopBarIconButton, TopBarIconButtonV2, TopBarTextButton } from 'src/navigator/TopBarButton'
+import { TopBarIconButton, TopBarTextButton } from 'src/navigator/TopBarButton'
 const testID = 'button'
 
 describe('TopBarTextButton', () => {
@@ -39,31 +39,5 @@ describe('TopBarIconButton', () => {
     )
     fireEvent.press(queryByTestId(testID)!)
     expect(onPress).toBeCalled()
-  })
-})
-
-describe('TopBarIconButtonV2', () => {
-  it('renders correctly with given props', () => {
-    const onPress = jest.fn()
-    const { getByTestId } = render(
-      <TopBarIconButtonV2 testID={testID} icon={<Text>IconV2</Text>} onPress={onPress} />
-    )
-    expect(getByTestId(testID)).toBeTruthy()
-    fireEvent.press(getByTestId(testID))
-    expect(onPress).toHaveBeenCalled()
-  })
-
-  it('handles the disabled state correctly', () => {
-    const onPress = jest.fn()
-    const { getByTestId } = render(
-      <TopBarIconButtonV2
-        testID={`${testID}-disabled`}
-        icon={<Text>IconV2</Text>}
-        onPress={onPress}
-        disabled={true}
-      />
-    )
-    fireEvent.press(getByTestId(`${testID}-disabled`))
-    expect(onPress).not.toHaveBeenCalled()
   })
 })
