+++ src/navigator/TopBarIconButtonV2.tsx
@@ -1,10 +1,9 @@
 import * as React from 'react'
-import { StyleProp, StyleSheet, Text, TextStyle, ViewStyle } from 'react-native'
+import { StyleProp, StyleSheet, View, ViewStyle } from 'react-native'
 import { AnalyticsEventType, AnalyticsPropertiesList } from 'src/analytics/Properties'
 import ValoraAnalytics from 'src/analytics/ValoraAnalytics'
 import Touchable from 'src/components/Touchable'
-import colors from 'src/styles/colors'
-import fontStyles from 'src/styles/fonts'
+import { Spacing } from 'src/styles/styles'
 import variables from 'src/styles/variables'
 
 interface CommonProps {
@@ -57,32 +56,29 @@ export type TopBarIconButtonProps = CommonProps & {
   containerStyle?: ViewStyle
 }
 
-/**
- * Please avoid use in new header icons
- * TODO: ACT-1134 alternative component
- * @deprecated
- */
-export function TopBarIconButton(props: TopBarIconButtonProps) {
-  return <Wrapper {...props}>{props.icon}</Wrapper>
-}
-
-export type TopBarTextButtonProps = CommonProps & {
-  title: string
-  titleStyle?: StyleProp<TextStyle>
-}
-
-export function TopBarTextButton(props: TopBarTextButtonProps) {
-  const { titleStyle, title } = props
+export function TopBarIconButtonV2(props: TopBarIconButtonProps) {
   return (
     <Wrapper {...props}>
-      <Text style={titleStyle ? [styles.text, titleStyle] : styles.text}>{title}</Text>
+      <View style={[styles.container, props.containerStyle]}>
+        <Touchable
+          testID={props.testID}
+          onPress={props.onPress}
+          style={[styles.button, props.style]}
+          borderRadius={Spacing.Thick24}
+        >
+          {props.icon}
+        </Touchable>
+      </View>
     </Wrapper>
   )
 }
 
 const styles = StyleSheet.create({
-  text: {
-    ...fontStyles.regular,
-    color: colors.primary,
+  button: {
+    padding: Spacing.Small12,
+  },
+  container: {
+    justifyContent: 'center',
+    alignItems: 'center',
   },
 })
